+++ src/virtual-wan/azext_vwan/_client_factory.py
@@ -18,12 +18,19 @@ def network_client_factory_2021_03_01(cli_ctx, aux_subscriptions=None, **_):
                                    api_version='2021-03-01')
 
 
+def network_client_factory_2021_08_01(cli_ctx, aux_subscriptions=None, **_):
+    from azure.cli.core.commands.client_factory import get_mgmt_service_client
+    from .profiles import CUSTOM_VWAN_2021_08_01
+    return get_mgmt_service_client(cli_ctx, CUSTOM_VWAN_2021_08_01, aux_subscriptions=aux_subscriptions,
+                                   api_version='2021-08-01')
+
+
 def cf_virtual_wans(cli_ctx, _):
     return network_client_factory(cli_ctx).virtual_wans
 
 
 def cf_virtual_hubs(cli_ctx, _):
-    return network_client_factory(cli_ctx).virtual_hubs
+    return network_client_factory_2021_08_01(cli_ctx).virtual_hubs
 
 
 def cf_virtual_hub_connection(cli_ctx, _):