+++ src/virtual-wan/azext_vwan/_params.py
@@ -54,6 +54,7 @@ def load_arguments(self, _):
         c.argument('virtual_wan', options_list='--vwan', help='Name or ID of the virtual WAN.', validator=get_network_resource_name_or_id('virtual_wan', 'virtualWans'))
         c.argument('address_prefix', help='CIDR address prefix for the virtual hub.')
         c.argument('sku', arg_type=get_enum_type(['Basic', 'Standard']), help='The sku of the VirtualHub.')
+        c.argument('hub_routing_preference', arg_type=get_enum_type(['ExpressRoute', 'VpnGateway', 'ASPath']), help='The routing preference in virtual hub of customer.')
 
     with self.argument_context('network vhub', arg_group='Gateway') as c:
         c.argument('express_route_gateway', help='Name or ID of an ExpressRoute gateway.', validator=get_network_resource_name_or_id('express_route_gateway', 'expressRouteGateways'))