+++ src/virtual-wan/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.2.12
+++++++
+*`az network vhub create`:support new parameters `--hub-routing-preference`.
+*`az network vhub update`:support new parameters `--hub-routing-preference`.
+
 0.2.11
 ++++++
 * add new command group: `az network vpn-gateway connection vpn-site-link-conn`.